+++ src/Encamina.Enmarcha.SemanticKernel.Connectors.Document/Exceptions/DocumentTooLargeException.cs
@@ -1,6 +1,4 @@
-﻿using System.Runtime.Serialization;
-
-namespace Encamina.Enmarcha.SemanticKernel.Connectors.Document.Exceptions;
+﻿namespace Encamina.Enmarcha.SemanticKernel.Connectors.Document.Exceptions;
 
 /// <summary>
 /// The exception that is thrown when there has been an error with the document size.
@@ -33,15 +31,6 @@ public class DocumentTooLargeException : Exception
     /// current exception is raised in a catch block that handles the inner exception.
     /// </param>
     public DocumentTooLargeException(string message, Exception innerException) : base(message, innerException)
-    {
-    }
-
-    /// <summary>
-    /// Initializes a new instance of the <see cref="DocumentTooLargeException"/> class.
-    /// </summary>
-    /// <param name="info">The object that holds the serialized object data.</param>
-    /// <param name="context">The contextual information about the source or destination.</param>
-    protected DocumentTooLargeException(SerializationInfo info, StreamingContext context) : base(info, context)
     {
     }
 }